+++ querydsl-spatial/src/main/java/com/querydsl/spatial/AbstractGeometryCollectionExpression.java
@@ -13,9 +13,12 @@
  */
 package com.querydsl.spatial;
 
+import com.querydsl.core.types.ConstantImpl;
 import com.querydsl.core.types.Expression;
+import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.core.types.dsl.NumberExpression;
-import org.geolatte.geom.GeometryCollection;
+import org.geolatte.geom.AbstractGeometryCollection;
+import org.geolatte.geom.Geometry;
 import org.jetbrains.annotations.Nullable;
 
 /**
@@ -25,15 +28,37 @@ import org.jetbrains.annotations.Nullable;
  *
  * @param <T>
  */
-public abstract class GeometryCollectionExpression<T extends GeometryCollection> extends AbstractGeometryCollectionExpression<T> {
+public abstract class AbstractGeometryCollectionExpression<T extends AbstractGeometryCollection> extends GeometryExpression<T> {
 
     private static final long serialVersionUID = 8874174644259834690L;
 
     @Nullable
     private transient volatile NumberExpression<Integer> numGeometries;
 
-    public GeometryCollectionExpression(Expression<T> mixin) {
+    public AbstractGeometryCollectionExpression(Expression<T> mixin) {
         super(mixin);
+    }
+
+    /**
+     * Returns the number of geometries in this GeometryCollection.
+     *
+     * @return number of geometries
+     */
+    public NumberExpression<Integer> numGeometries() {
+        if (numGeometries == null) {
+            numGeometries = Expressions.numberOperation(Integer.class, SpatialOps.NUM_GEOMETRIES, mixin);
+        }
+        return numGeometries;
+    }
+
+    /**
+     * Returns the Nth geometry in this GeometryCollection.
+     *
+     * @param n one based index
+     * @return matching geometry
+     */
+    public GeometryExpression<Geometry> geometryN(Integer n) {
+        return GeometryExpressions.geometryOperation(SpatialOps.GEOMETRYN, mixin, ConstantImpl.create(n));
     }
 
 }